+++ CHANGES.rst
@@ -20,6 +20,8 @@ Development version
 * Close :issue:`112`: Implement a ``--report=minimal`` option for printing
   a succinct two-line report in tab-separated value (tsv) format. Thanks
   to :user:`jvolkening` for coming up with an initial patch!
+* The options ``--no-trim`` and ``--mask-adapter`` should now be written as
+  ``--action=mask`` and ``--action=none``. The old options still work.
 * Fix :issue:`128`: The “Reads written” figure in the report incorrectly
   included both trimmed and untrimmed reads if ``--untrimmed-output`` was used.
 * This is the last release to support :ref:`colorspace data <colorspace>`.